+++ docker-compose.yml
@@ -80,6 +80,7 @@ services:
       - ./services/madoc-ts/service-jwts:/home/node/app/service-jwts
       - ./services/madoc-ts/themes:/home/node/app/themes
       - ./services/madoc-ts/translations:/home/node/app/translations
+      - ./services/madoc-ts/npm:/home/node/app/npm
       - ./var/certs:/openssl-certs
       - ./var/jwt:/home/node/app/service-jwt-responses
       - ./var/files:/home/node/app/omeka-files
