+++ services/madoc-ts/ecosystem.config.js
@@ -5,7 +5,7 @@ module.exports = {
       ? [
           {
             name: 'server',
-            script: 'lib/index.js',
+            script: 'lib/server.js',
             instances: 1,
             autorestart: true,
             watch: false,
@@ -40,7 +40,7 @@ module.exports = {
       : [
           {
             name: 'server',
-            script: 'lib/index.js',
+            script: 'lib/server.js',
             instances: 1,
             autorestart: true,
             watch: ['lib'],
