+++ services/madoc-ts/npm/madoc-types/.gitignore
@@ -1,0 +1,1 @@
+dist
