+++ services/madoc-ts/npm/madoc-types/index.js
@@ -1,0 +1,1 @@
+module.exports = typeof window !== 'undefined' ? window.Madoc : global.Madoc;
