+++ services/madoc-ts/src/frontend/admin/pages/content/linking/view-ocr-model.tsx
@@ -3,8 +3,8 @@ import { CaptureModel } from '@capture-models/types';
 import * as React from 'react';
 import { Suspense, useMemo } from 'react';
 import { useQuery } from 'react-query';
-import { ResourceLinkResponse } from '../../../../../database/queries/linking-queries';
 import { preprocessCaptureModel } from '../../../../../extensions/capture-models/Paragraphs/Paragraphs.helpers';
+import { ResourceLinkResponse } from '../../../../../types/schemas/linking';
 import { RevisionNavigation } from '../../../../shared/caputre-models/RevisionNavigation';
 import { documentFragmentWrapper } from '../../../../shared/caputre-models/utility/document-fragment-wrapper';
 import { useApi } from '../../../../shared/hooks/use-api';
