+++ services/madoc-ts/src/frontend/shared/plugins/external/server-renderer-for.ts
@@ -2,7 +2,7 @@ import React from 'react';
 import { ApiClient } from '../../../../gateway/api';
 
 export function serverRendererFor<TVariables = any, Data = any>(
-  component: React.FC<Data>,
+  component: React.FC<any>,
   config: {
     getKey?: (params: any, query: any, pathname: string) => [string, TVariables];
     getData?: (key: string, vars: TVariables, api: ApiClient, pathname: string) => Promise<Data>;
