+++ services/madoc-ts/src/frontend/shared/plugins/public-api.ts
@@ -1,0 +1,22 @@
+import { useCollectionList } from '../../site/hooks/use-collection-list';
+import { Button } from '../atoms/Button';
+import { LocaleString } from '../components/LocaleString';
+import { useApi } from '../hooks/use-api';
+import { blockConfigFor } from './external/block-config-for';
+import { serverRendererFor } from './external/server-renderer-for';
+import { atoms, useAtoms } from './use-atoms';
+import { useComponents } from './use-components';
+import { useModule } from './use-module';
+
+export {
+  useModule as require,
+  useApi,
+  atoms as Atoms,
+  Button,
+  LocaleString,
+  useAtoms,
+  useComponents,
+  useCollectionList,
+  blockConfigFor,
+  serverRendererFor,
+};
